# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/member/products/{productId}/draws": {
      "get": {
        "operationId": "MemberDrawController_listDraws",
        "parameters": [
          {
            "name": "state",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "DRAFT",
                "SCHEDULED",
                "OPEN",
                "CLOSED",
                "RESULT_PENDING",
                "RESULT_CONFIRMED",
                "SETTLING",
                "SETTLED"
              ],
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberDrawPageBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List Lottery Draws for a Product (betting discovery)",
        "tags": [
          "Member Lottery Draws"
        ]
      }
    },
    "/api/v1/member/draws/{id}": {
      "get": {
        "operationId": "MemberDrawController_getDraw",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberDrawDetailBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a Lottery Draw detail with status, cutoff and server time",
        "tags": [
          "Member Lottery Draws"
        ]
      }
    },
    "/api/v1/member/draws/{id}/eligibility": {
      "get": {
        "operationId": "MemberDrawController_eligibility",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberDrawEligibilityBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cutoff eligibility: strictly-before eligible; exact/beyond rejected",
        "tags": [
          "Member Lottery Draws"
        ]
      }
    },
...
  },
  "info": {
    "title": "Lottify v1 API",
    "description": "Authoritative REST/OpenAPI contract for Lottify v1",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "MemberDrawCutoffBody": {
        "type": "object",
        "properties": {
          "cutoffAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "cutoffAt"
        ]
      },
      "MemberDrawBetTypeBody": {
        "type": "object",
        "properties": {
          "betTypeId": {
            "type": "string"
          },
          "betTypeCode": {
            "type": "string"
          },
          "betTypeVersionId": {
            "type": "string"
          },
          "canonicalNumberFormat": {
            "type": "string"
          },
          "validationPattern": {
            "type": "string"
          },
          "payout": {
            "type": "object",
            "description": "Resolved payout configuration (opaque configuration value)."
          },
          "minStakeMinor": {
            "type": "string",
            "description": "Minimum stake in minor units."
          },
          "maxStakeMinor": {
            "type": "string",
            "description": "Maximum stake in minor units."
          },
          "limitPolicyRef": {
            "type": "string"
          },
          "restrictionPolicyRef": {
            "type": "string"
          },
          "settlementRuleVersionRef": {
            "type": "string"
          }
        },
        "required": [
          "betTypeId",
          "betTypeCode",
          "betTypeVersionId",
          "canonicalNumberFormat",
          "validationPattern",
          "payout",
          "minStakeMinor",
          "maxStakeMinor",
          "limitPolicyRef",
          "restrictionPolicyRef",
          "settlementRuleVersionRef"
        ]
      },
      "MemberDrawDetailBody": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "productId": {
            "type": "string"
          },
          "productVersionId": {
            "type": "string"
          },
          "occurrenceIdentity": {
            "type": "string"
          },
          "localDate": {
            "type": "string"
          },
          "state": {
            "type": "string",
            "enum": [
              "DRAFT",
              "SCHEDULED",
              "OPEN",
              "CLOSED",
              "RESULT_PENDING",
              "RESULT_CONFIRMED",
              "SETTLING",
              "SETTLED",
              "CANCELLING",
              "CANCELLED"
            ]
          },
          "version": {
            "type": "number"
          },
          "openAt": {
            "type": "string",
            "format": "date-time"
          },
          "cutoffAt": {
            "type": "string",
            "format": "date-time"
          },
          "drawAt": {
            "type": "string",
            "format": "date-time"
          },
          "provenance": {
            "type": "string"
          },
          "timezone": {
            "type": "string"
          },
          "scheduleTemplateRef": {
            "type": "string"
          },
          "resultSchemaVersionRef": {
            "type": "string"
          },
          "settlementRuleVersionRef": {
            "type": "string"
          },
          "defaultPayoutPolicyRef": {
            "type": "string"
          },
          "defaultLimitPolicyRef": {
            "type": "string"
          },
          "defaultRestrictionPolicyRef": {
            "type": "string"
          },
          "resultSourceRef": {
            "type": "string",
            "nullable": true
          },
          "overrideRevisionRef": {
            "type": "string"
          },
          "cutoff": {
            "$ref": "#/components/schemas/MemberDrawCutoffBody"
          },
          "serverNow": {
            "type": "string",
            "format": "date-time"
          },
          "allowedActions": {
            "description": "Lifecycle commands currently allowed by the Draw state machine.",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "betTypes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MemberDrawBetTypeBody"
            }
          }
        },
        "required": [
          "id",
          "productId",
          "productVersionId",
          "occurrenceIdentity",
          "localDate",
          "state",
          "version",
          "openAt",
          "cutoffAt",
          "drawAt",
          "provenance",
          "timezone",
          "scheduleTemplateRef",
          "resultSchemaVersionRef",
          "settlementRuleVersionRef",
          "defaultPayoutPolicyRef",
          "defaultLimitPolicyRef",
          "defaultRestrictionPolicyRef",
          "resultSourceRef",
          "overrideRevisionRef",
          "cutoff",
          "serverNow",
          "allowedActions",
          "betTypes"
        ]
      },
      "MemberDrawPageBody": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MemberDrawDetailBody"
            }
          },
          "nextCursor": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "items",
          "nextCursor"
        ]
      },
      "MemberDrawEligibilityBody": {
        "type": "object",
        "properties": {
          "eligible": {
            "type": "boolean",
            "description": "True only when server time is strictly before the Draw cutoff."
          },
          "cutoffAt": {
            "type": "string",
            "format": "date-time"
          },
          "serverNow": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "eligible",
          "cutoffAt",
          "serverNow"
        ]
      },
...
    }
  }
}